Add files via upload
I have built a Flight Booking Agency with the name of "FLYHIGH". The motive of my project is to build a flight booking system which allows you to book tickets for upcoming flights to your dream destination. The project displays your ticket at the end with all the necessary details. I have used basic Java programming to complete the task!!!